InvPerfWebApi/InvPerfWebApi.Tests/Controllers/HomeControllerTest.cs
using InvPerfWebApi;
using InvPerfWebApi.Controllers;
using Microsoft.VisualStudio.TestTools.UnitTesting;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Web.Mvc;

namespace InvPerfWebApi.Tests.Controllers
{
    [TestClass]
    public class HomeControllerTest
    {
        [TestMethod]
        public void Index()
        {
            // Arrange
            HomeController controller = new HomeController();

            // Act
            ViewResult result = controller.Index() as ViewResult;

            // Assert
            Assert.IsNotNull(result);
        }

        [TestMethod]
        public void About()
        {
            // Arrange
            HomeController controller = new HomeController();

            // Act
            ViewResult result = controller.About() as ViewResult;

            // Assert
            Assert.AreEqual("Your application description page.", result.ViewBag.Message);
        }

        [TestMethod]
        public void Contact()
        {
            // Arrange
            HomeController controller = new HomeController();

            // Act
            ViewResult result = controller.Contact() as ViewResult;

            // Assert
            Assert.IsNotNull(result);
        }

        [TestMethod]
        public async void GetGetUserInvestmentDataAsync()
        {
            //Testing the Investment data view 
            // Arrange
            HomeController controller = new HomeController();

            string userName = "User4";
            string investmentName = "Investment4";
            // Act
            ViewResult result = await controller.GetUserInvestmentDataAsync(userName, investmentName) as ViewResult;

            // Assert
            Assert.IsNotNull(result);
        }
    }
}
